+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Added
+
+- Support importing remote portable actor profiles.
+
 ### Changed
 
 - Accept `Update(Person)` C2S activities.